requirements.txt
click==8.1.7
exceptiongroup==1.2.0
gitdb==4.0.11
GitPython==3.1.42
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.3
smmap==5.0.1
tomli==2.0.1
semantic-version==2.10.0